refactor the exhttp package
