+++ exhttp/go.mod
@@ -1,0 +1,15 @@
+module github.com/hasura/ndc-http/exhttp
+
+go 1.23.0
+
+require github.com/hasura/ndc-sdk-go v1.7.0
+
+require (
+	github.com/go-viper/mapstructure/v2 v2.2.1 // indirect
+	github.com/google/uuid v1.6.0 // indirect
+	github.com/prometheus/client_model v0.6.1 // indirect
+	github.com/prometheus/common v0.61.0 // indirect
+	go.opentelemetry.io/otel v1.33.0 // indirect
+	google.golang.org/protobuf v1.36.0 // indirect
+	gotest.tools/v3 v3.5.2 // indirect
+)
