+++ go.work
@@ -3,4 +3,5 @@ go 1.23.0
 use (
 	.
 	./ndc-http-schema
+	./exhttp
 )
