+++ ndc-http-schema/go.mod
@@ -5,17 +5,18 @@ go 1.23.0
 toolchain go1.23.4
 
 require (
-	github.com/alecthomas/kong v1.6.0
+	github.com/alecthomas/kong v1.8.1
 	github.com/evanphx/json-patch v0.5.2
 	github.com/google/go-cmp v0.6.0
-	github.com/hasura/ndc-sdk-go v1.6.4-0.20241220173928-1c66c55ba78d
-	github.com/invopop/jsonschema v0.12.0
-	github.com/lmittmann/tint v1.0.6
+	github.com/hasura/ndc-http/exhttp v0.0.1
+	github.com/hasura/ndc-sdk-go v1.7.0
+	github.com/invopop/jsonschema v0.13.0
+	github.com/lmittmann/tint v1.0.7
 	github.com/pb33f/libopenapi v0.18.7
 	github.com/theory/jsonpath v0.2.1
 	github.com/wk8/go-ordered-map/v2 v2.1.9-0.20240815153524-6ea36470d1bd
 	gopkg.in/yaml.v3 v3.0.1
-	gotest.tools/v3 v3.5.1
+	gotest.tools/v3 v3.5.2
 )
 
 require (
@@ -33,3 +34,5 @@ require (
 	go.opentelemetry.io/otel v1.33.0 // indirect
 	google.golang.org/protobuf v1.36.0 // indirect
 )
+
+replace github.com/hasura/ndc-http/exhttp => ../exhttp
