+++ ndc-http-schema/jsonschema/convert-config.schema.json
@@ -71,7 +71,7 @@
       "required": [
         "file"
       ],
-      "description": "ConvertConfig represents the content of convert config file"
+      "description": "ConvertConfig represents the content of convert config file."
     },
     "PatchConfig": {
       "properties": {
